Use parse-body from alexandria instead of uiop
UIOP:PARSE-BODY is available from uiop past version 3.2.0; some
implementations (most notably ECL) frozen ASDF at the version prior to this
release.

Technically it is possible to put (:version "uiop" "3.2.0") as a dependency in
the system, although Quicklisp won't download it because ASDF is able to locate
UIOP system, so Quicklisp does not recognize it as a missing dependency. This
means that the user is forced to manually download uiop.

Alexandria is already dependency of CFFI, so this commit does not add new
dependencies.